user is now added to state for authorised routes

+++ server/recipe-app/routes/controllers/recipeController.ts
@@ -16,7 +16,7 @@ const getRecipeData = async (ctx: Context) => {
   return {
     title: jsonBody.title,
     description: jsonBody.description,
-    ingredients: jsonBody.ingredients
+    ingredients: jsonBody.ingredients,
   };
 };
 
@@ -55,7 +55,7 @@ const addRecipe = async (ctx: Context) => {
     recipeValidationRules,
   );
   
-  const recipe = await recipeService.findRecipeByUserIdAndTitle(1, recipeData.title);
+  const recipe = await recipeService.findRecipeByUserIdAndTitle(ctx.state.user.id, recipeData.title);
 
   if (recipe !== -1) {
     console.error("Unique key constraint violation!");
@@ -71,12 +71,12 @@ const addRecipe = async (ctx: Context) => {
     return;
   } else {
     await recipeService.addRecipe(
-      1,
+      ctx.state.user.id,
       recipeData.title,
       recipeData.description
     );
 
-    const recipe = await recipeService.findRecipeByUserIdAndTitle(1, recipeData.title);
+    const recipe = await recipeService.findRecipeByUserIdAndTitle(ctx.state.user.id, recipeData.title);
 
     if (recipe === -1) {
       console.error("Could not create recipe. Please try again.");
@@ -107,6 +107,13 @@ const deleteRecipe = async (ctx: any) => {
     return;
   }
 
+  if (recipe.user_id != ctx.state.user.id) {
+    console.error("Not authorised to delete recipe.");
+    ctx.response.status = 401; // Bad Request
+    ctx.response.body = "Not authorised to delete this recipe!";
+    return;
+  }
+
   await recipeService.deleteRecipe(rId);
 
   ctx.response.status = 200; // OK
@@ -129,12 +136,19 @@ const updateRecipe = async (ctx: any) => {
     return;
   }
 
-  const recipe_duplicate = await recipeService.findRecipeByUserIdAndTitle(1, recipeData.title);
+  const recipe_duplicate = await recipeService.findRecipeByUserIdAndTitle(ctx.state.user.id, recipeData.title);
 
   if (recipe_duplicate !== -1) {
     console.error("Unique key constraint violation!");
     ctx.response.status = 400;
     ctx.response.body = "Recipe could not be created because a user can not have multiple recipes with the same name!";
+    return;
+  }
+
+  if (recipe.user_id != ctx.state.user.id) {
+    console.error("Not authorised to delete recipe.");
+    ctx.response.status = 401; // Bad Request
+    ctx.response.body = "Not authorised to delete this recipe!";
     return;
   }
 
@@ -152,7 +166,7 @@ const updateRecipe = async (ctx: any) => {
 
     await ingredientService.deleteIngredientsOfRecipeById(recipe.id);
 
-    recipeData.ingredients.forEach(async (name: any) => {
+    recipeData.ingredients.forEach(async (name: string) => {
       await ingredientService.addIngredient(
         recipe.id,
         name
